SPMI: Hacky workaround for coredistools unsupported instructions (#91668)
We currently are unable to get SPMI diffs for methods that use ldapur
instructions as our current coredistools does not support this
instruction. While we wait for coredistools to be updated (this was
blocked on some C++ lib issues) here is a hacky temporary workaround for
the problem. It rewrites ldapur* instructions into the corresponding
ldur* instruction before passing it to the near differ.