Fix type parameter mapping logic in ILLink/ILCompiler (#92088)
The compiler-generated code for `Container<T>` in this testcase
contains two closure types, one for each lambda in the expression
`(Func<T, T> x) => v => x(v);`.

The outer closure environment type is instantiated by its cctor,
which stores an instance into a static field. It is never
directly instantiated from another type - instead, the cctor is
triggered when the cctor of `Container<T>` access this static
field.

This case was missing from the compiler-generated type parameter
mapping logic. We need to discover the substituted generic
typeref for the outer closure environment from the `ldsfld`.