[mono][interp] Fix calling of static delegates with simd arguments (#92927)
* [mono][interp] Small refactoring to make the code clearer

Previous code was checking if we should have already pushed a dreg, to compute the correct position of the argument on the stack. Save pointer to args directly instead.

* [mono][interp] Fix calling of static delegates with simd arguments

If we do a delegate call that translates to a static call, the delegate pointer that was passed needs to be skipped and we need to pass the actual arguments. However, the following arguments might not come immediately after the delegate pointer, if the first argument is simd aligned. Pass this information to MINT_CALL_DELEGATE so the correct offset of the arguments can be computed.

* [mono][interp] Don't pass csignature to MINT_CALL_DELEGATE

Pass param_count directly