Fix passing CMAKE_BUILD_TYPE in runtime tests using Ninja (#93051)
I noticed that we were building the native assets of runtime tests always in Debug mode on Windows when using Ninja, even though we passed CMAKE_BUILD_TYPE=Release.

This results in warnings like this about mismatching debug/release msvc runtime when building with NativeAOT (which uses release msvc runtime in Release config):

> [warning]LIBCMT.lib(initializers.obj)(0,0): warning LNK4098: defaultlib 'libcmtd.lib' conflicts with use of other libs; use /NODEFAULTLIB:library

It turned out that this is because we were accidentally passing CMAKE_SYSTEM_VERSION and CMAKE_BUILD_TYPE as a single string, instead of multiple. This resulted in the CMAKE_SYSTEM_VERSION being set like this in CMakeCache.txt:

```
CMAKE_SYSTEM_VERSION:UNINITIALIZED=10.0 -DCMAKE_BUILD_TYPE=Release
```

CMake then defaulted CMAKE_BUILD_TYPE to Debug.

This uncovered an MSVC bug when compiling some sources with the MSVC version used on CI in /O2 mode. This is fixed in 19.38+, disabled optimization in these files on older versions to workaround.